chore: move core handler logic
