+++ packages/core/src/events/index.tsx
@@ -1,3 +1,4 @@
 export { useEventHandler } from './EventContext';
-export { DerivedEventHandlers, EventHandlers } from './EventHandlers';
+export * from './CoreEventHandlers';
+export { EventHandlers } from './EventHandlers';
 export { Events } from './Events';
